+++ pages/chat.js
@@ -1,0 +1,11 @@
+
+export default function ChatRoom() {
+
+    return (
+        <>
+            <div>
+                ChatRoom
+            </div>
+        </>
+    )
+}
