+++ pages/index.js
@@ -1,35 +1,8 @@
 import appConfig from '../config.json';
 import { Box, Button, Text, TextField, Image } from '@skynexui/components';
+import React from 'react';
+import {useRouter} from 'next/router';
 
-
-function GlobalStyle() {
-    return (
-        <style global jsx>{`
-        * {
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-        list-style: none;
-        }
-        body {
-            font-family: 'Open Sans', sans-serif;
-        }
-        /* App fit Height */ 
-        html, body, #__next {
-            min-height: 100vh;
-            display: flex;
-            flex: 1;
-        }
-        #__next {
-            flex: 1;
-        }
-        #__next > * {
-            flex: 1;
-        }
-        /* ./App fit Height */ 
-        `}</style>
-    );
-}
 
 //Criação de tag
 function Title(props) {
@@ -69,12 +42,14 @@ function Title(props) {
 
 export default HomePage*/
 
-export default function PaginaInicial() {
-    const username = 'fgalletti93';
+export default function HomePage() {
+    const [username, setUserName] = React.useState('fgalletti93');
+    const routing = useRouter();
+
+
 
     return (
         <>
-            <GlobalStyle />
             <Box
                 styleSheet={{
                     display: 'flex', alignItems: 'center', justifyContent: 'center',
@@ -101,6 +76,10 @@ export default function PaginaInicial() {
                     {/* Formulário */}
                     <Box
                         as="form"
+                        onSubmit={function (eventInfo) {
+                            eventInfo.preventDefault();
+                            routing.push('/chat');
+                        }}
                         styleSheet={{
                             display: 'flex', flexDirection: 'column', alignItems: 'center', justifyContent: 'center',
                             width: { xs: '100%', sm: '50%' }, textAlign: 'center', marginBottom: '32px',
@@ -112,6 +91,14 @@ export default function PaginaInicial() {
                         </Text>
 
                         <TextField
+                            value={username}
+                            onChange={function (event) {
+                                //Onde ta o valor?
+                                const valor = event.target.value;
+                                //Trocar o valor da variavel
+                                //através do React e avise quem precisa
+                                setUserName(valor);
+                            }}
                             fullWidth
                             textFieldColors={{
                                 neutral: {
@@ -121,10 +108,10 @@ export default function PaginaInicial() {
                                     backgroundColor: appConfig.theme.colors.neutrals[800],
                                 },
                             }}
-                        />
+                        /> 
                         <Button
                             type='submit'
-                            label='Entrar'
+                            label='Unplug'
                             fullWidth
                             buttonColors={{
                                 contrastColor: appConfig.theme.colors.neutrals["000"],
